+++ lib/helpers/metadata_helpers.dart
@@ -32,4 +32,5 @@ class MetaDataHelper<T, DM> {
         return new MetaDataValue<T>(am1.reflectee, dm, dm.simpleName, instanceMirror);
       }));
   }
+  
 }