+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 This file contains highlights of what changes on each version of the forcemvc package. 
 
-#### Pub version 0.2.1+1 ####
+#### Pub version 0.2.1+1 & 0.2.1+2 ####
 
-add a method on MetadataValue to get declartion of, named dclOnMetaData
+add a method on MetadataValue to get declartion of, named dclWithMetaData
 
 #### Pub version 0.2.1 ####