+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file contains highlights of what changes on each version of the forcemvc package. 
 
-#### Pub version 0.2.1+1 & 0.2.1+2 & 0.2.1+3 ####
+#### Pub version 0.2.1+1 - 0.2.1+5 ####
 
 add a method on MetadataValue to get the return type typeOfOwner